refactoring: fix namespaces
