+++ tools/Evm/T8n/Errors/GethErrorMappings.cs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-namespace Evm.t8n.Errors;
+namespace Evm.T8n.Errors;
 
 public class GethErrorMappings
 {
