+++ tools/Evm/T8n/JsonTypes/RejectedTx.cs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-namespace Evm.t8n;
+namespace Evm.T8n.JsonTypes;
 
 public class RejectedTx(int index, string error)
 {
